Refactor:add custom hook & delete useEffect
